requirements.txt
asgiref==3.4.1
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
importlib-metadata==4.10.0
Markdown==3.3.6
Pillow==9.0.0
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
zipp==3.7.0
